requirements.txt
aiohttp>=3.10.0
rich>=13.7.0
PyYAML>=6.0.1
typer>=0.12.3
fastapi>=0.110.0
uvicorn>=0.28.0
websockets>=13.0.0
